requirements.txt
opencv-python
pillow
pygame
ultralytics